+++ lib/simulation/EnergyPlan.cpp
@@ -3,8 +3,8 @@
 namespace simulation {
 namespace config {
 
-int EnergyPlan::getEnergyFromWattage(int wattage) {
-  return wattage/60; //TODO 60: (unit is one minute) define somewhere (config?)
+double EnergyPlan::getEnergyFromWattage(double wattage) {
+  return wattage/60.0; //TODO 60: (unit is one minute) define somewhere (config?)
 }
 
 } /* End of namespace simulation.config */
