+++ lib/simulation/Light.h
@@ -11,7 +11,7 @@ class Light : public Consumer {
 
 public:
   Light(std::string consumerId) : Consumer(consumerId) {
-    int energy = config::EnergyPlan::getEnergyFromWattage(1000); //realistisch 200
+    double energy = config::EnergyPlan::getEnergyFromWattage(1000); //realistisch 200
 
     addEnergyPlan(new config::EnergyPlanStatic(390, 450, energy));
     addEnergyPlan(new config::EnergyPlanStatic(1020, 1350, energy));
